Crate op_alloy_network

op-alloy-network

Optimism blockchain RPC behavior abstraction.

This crate contains a simple abstraction of the RPC behavior of an Op-stack blockchain. It is intended to be used by the Alloy client to provide a consistent interface to the rest of the library, regardless of changes the underlying blockchain makes to the RPC interface.

Modules

eip2718

EIP-2718 traits.

primitives

alloy-network-primitives

Structs

AnyNetwork

Types for a catch-all network.

AnyTxType

Transaction type for a catch-all network.

Ethereum

Types for a mainnet-like Ethereum network.

EthereumWallet

A wallet capable of signing any transaction for the Ethereum network.

Optimism

Types for an Op-stack network.

UnbuiltTransactionError

An unbuilt transaction, along with some error.

Enums

TransactionBuilderError

Error type for transaction builders.

Traits

Network

Captures type info for network-specific RPC requests/responses.

NetworkWallet

A wallet capable of signing any transaction for the given network.

ReceiptResponse

Receipt JSON-RPC response.

TransactionBuilder

A Transaction builder for a network.

TransactionResponse

Transaction JSON-RPC response.

TxSigner

Asynchronous transaction signer, capable of signing any [SignableTransaction] for the given Signature type.

TxSignerSync

Synchronous transaction signer, capable of signing any [SignableTransaction] for the given Signature type.

Type Aliases

BuildResult

Result type for transaction builders