Struct op_alloy_network::EthereumWallet

pub struct EthereumWallet {
    default: Address,
    signers: BTreeMap<Address, Arc<dyn TxSigner<Signature<Signature<Secp256k1>>> + Sync + Send>>,
}

A wallet capable of signing any transaction for the Ethereum network.

Fields
default: Addresssigners: BTreeMap<Address, Arc<dyn TxSigner<Signature<Signature<Secp256k1>>> + Sync + Send>>
Implementations

impl EthereumWallet

pub fn new<S>(signer: S) -> EthereumWallet
where S: TxSigner<Signature<Signature<Secp256k1>>> + Send + Sync + 'static,

Create a new signer with the given signer as the default signer.

pub fn register_signer<S>(&mut self, signer: S)
where S: TxSigner<Signature<Signature<Secp256k1>>> + Send + Sync + 'static,

Register a new signer on this object. This signer will be used to sign TransactionRequest and [TypedTransaction] object that specify the signer’s address in the from field.

pub fn register_default_signer<S>(&mut self, signer: S)
where S: TxSigner<Signature<Signature<Secp256k1>>> + Send + Sync + 'static,

Register a new signer on this object, and set it as the default signer. This signer will be used to sign TransactionRequest and [TypedTransaction] objects that do not specify a signer address in the from field.

pub fn default_signer( &self, ) -> Arc<dyn TxSigner<Signature<Signature<Secp256k1>>> + Sync + Send>

Get the default signer.

pub fn signer_by_address( &self, address: Address, ) -> Option<Arc<dyn TxSigner<Signature<Signature<Secp256k1>>> + Sync + Send>>

Get the signer for the given address.